QAP 4 - WebDesign - JavaScript
Project 5 – JavaScript
Create a reasonably complicated Object in JavaScript that contains attributes and methods for the Motel
customer. The customer attributes should include (and not be limited to), the customer’s name, birth
date, gender, and room preferences (as an array), payment method, mailing address (as a sub-object),
phone number, check-in, and check-out date (as a sub-object), and object methods to determine their
age and duration of stay. The JavaScript code should also create a template literal string, or properly
formatted html, that describes the customer. In other words, writes a paragraph describing the customer
with many of their personal attributes dynamically embedded in the string / html coming from the above
object definition.